+++ deutsch/__init__.py
@@ -10,4 +10,4 @@ from .lexer import Lexer
 from .parser import Parser
 from .interpreter import Interpreter
 
-__version__ = '2.6.0'
+__version__ = '2.7.0'