+++ deutsch/lexer.py
@@ -31,6 +31,9 @@ SCHLUESSELWOERTER = {
     'fall':      TokenTyp.FALL,
     'werfe':     TokenTyp.WERFE,
     'pruefe':    TokenTyp.PRUEFE,
+    'konstante': TokenTyp.KONSTANTE,
+    'statisch':  TokenTyp.STATISCH,
+    'als':       TokenTyp.ALS,
 }
 
 _ESCAPES = {'n': '\n', 't': '\t', 'r': '\r', '\\': '\\',